Created Movement Script
I created a movement script which can be attached to a player. When it's the player's turn, a boolean variable isUsersTurn is set to true, and they're allowed to move for 5 seconds, and then isUsersTurn is set to false. isUsersTurn must be set to true again for the player to move. If the player is out, isPlayerOut is set to true. The speed of the player, the time they can move, and the booleans isUsersTurn and isPlayerOut are all accessible from outside the script (they're public variables).